# api/openapi.yaml
openapi: 3.0.3
info:
  title: IMS ERP System API
  description: |
    Complete ERP system API with microservices architecture
    
    ## Architecture
    - **API Gateway** (8080) - Central entry point for all client requests
    - **Auth Service** (8081) - Authentication and authorization
    - **Client Services** (8082) - Client management (CQRS pattern)
    - **Invoice Service** (8083) - Invoice management
    - **Payment Service** (8084) - Payment processing
    - **Product Service** (8085) - Product catalog
    - **Order Service** (8086) - Order management
    - **Inventory Service** (8087) - Inventory control
    
    ## Authentication
    All protected endpoints require Bearer token authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per tenant.
    
    ## Error Handling
    All errors follow standard HTTP status codes with detailed error messages.
  version: 1.0.0
  contact:
    name: IMS ERP Support
    email: support@ims-erp.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.ims-erp.example.com/api/v1
    description: Production server
...
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: List orders
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, pending, confirmed, processing, shipped, delivered, completed, cancelled, refunded]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of orders
    post:
      tags:
        - Orders
      summary: Create order
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  orderNumber:
                    type: string
...
components:
  schemas:
...
    CreateOrderRequest:
      type: object
      required:
        - clientId
        - type
        - source
      properties:
        clientId:
          type: string
          format: uuid
        type:
          type: string
          enum: [standard, pre_order, subscription, quote]
        source:
          type: string
          enum: [web, mobile, api, phone, in_store, marketplace]
        currency:
          type: string
          default: USD
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineRequest'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
    OrderLineRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
          default: 0
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []